requirements.txt
fastmcp==3.2.0
requests==2.32.5
uvicorn==0.42.0
python-dotenv==1.1.0
